# account/src/main/resources/api/api.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Account Service
  description: API for managing account accounts.
servers:
  - url: 'http:/localhost:8080/api'
paths:
  /account/all:
    get:
      tags:
        - Account
      summary: Getting all accounts
      operationId: getAllAccounts
      responses:
        "201":
          description: Successful retrieval of all accounts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /account/all/active:
    get:
      tags:
        - Account
      summary: Getting all active accounts
      operationId: getAllActiveAccounts
      responses:
        "201":
          description: Successful retrieval of all active accounts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /account/{id}:
    get:
      tags:
        - Account
      summary: Getting single account by id
      operationId: getAccountById
      responses:
        "201":
          description: Successful retrieval of account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: integer
            minimum: 1
          in: query
          name: Account
          required: true
          description: The account id
  /account/{id}/delete:
    get:
      tags:
        - Account
      summary: Deleting single account by id
      operationId: deleteAccount
      responses:
        "201":
          description: Successful deletion of account.
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: integer
            minimum: 1
          in: query
          name: Account
          required: true
          description: The account id
  /account/{email}:
    get:
      tags:
        - Account
      summary: Getting single account by email
      operationId: getAccountByEmail
      responses:
        "201":
          description: Successful retrieval of account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: string
          in: query
          name: Account
          required: true
          description: Unique account email
  /account/update:
    post:
      tags:
        - Account
      summary: Update account
      operationId: updateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountDTO"
      responses:
        "201":
          description: Successful update of account.
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /account/create:
    post:
      tags:
        - Account
      summary: Create account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountDTO"
      responses:
        "201":
          description: Successful creation of account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"

  /player/all:
    get:
      tags:
        - Player
      summary: Getting all players
      operationId: getAllPlayers
      responses:
        "201":
          description: Successful retrieval of all players.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /player/all/active:
    get:
      tags:
        - Player
      summary: Getting all active players
      operationId: getAllActivePlayers
      responses:
        "201":
          description: Successful retrieval of all players.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /player/{id}:
    get:
      tags:
        - Player
      summary: Getting single player by id
      operationId: getPlayerById
      responses:
        "201":
          description: Successful retrieval of player.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: integer
            minimum: 1
          in: query
          name: Player
          required: true
          description: The player id
  /player/{id}/deactivate:
    get:
      tags:
        - Player
      summary: Deactivating single player by id
      operationId: deactivatePlayer
      responses:
        "201":
          description: Successful deactivation of player.
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: integer
            minimum: 1
          in: query
          name: Player
          required: true
          description: The player id
  /player/{email}:
    get:
      tags:
        - Player
      summary: Getting single player by email
      operationId: getPlayerByEmail
      responses:
        "201":
          description: Successful retrieval of player.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: string
          in: query
          name: Player
          required: true
          description: Unique player email
  /player/update:
    post:
      tags:
        - Player
      summary: Update player
      operationId: updatePlayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerDTO"
      responses:
        "201":
          description: Successful update of player.
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /player/create:
    post:
      tags:
        - Player
      summary: Create player
      operationId: createPlayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlayerDTO'
      responses:
        "201":
          description: Successful creation of player.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDTO'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"

  /team/all:
    get:
      tags:
        - Team
      summary: Getting all teams
      operationId: getAllTeams
      responses:
        "201":
          description: Successful retrieval of all teams.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /team/all/active:
    get:
      tags:
        - Team
      summary: Getting all active teams
      operationId: getAllActiveTeams
      responses:
        "201":
          description: Successful retrieval of all teams.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /team/{id}:
    get:
      tags:
        - Team
      summary: Getting single team by id
      operationId: getTeamById
      responses:
        "201":
          description: Successful retrieval of team.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: integer
            minimum: 1
          in: query
          name: team
          required: true
          description: The team id
  /team/{id}/deactivate:
    get:
      tags:
        - Team
      summary: Deactivating single team by id
      operationId: deactivateTeam
      responses:
        "201":
          description: Successful deactivation of team.
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: integer
            minimum: 1
          in: query
          name: team
          required: true
          description: The team id
  /team/update:
    post:
      tags:
        - Team
      summary: Update team
      operationId: updateTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamDTO"
      responses:
        "201":
          description: Successful update of team.
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /team/create:
    post:
      tags:
        - Team
      summary: Create team
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamDTO'
      responses:
        "201":
          description: Successful creation of team.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDTO'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /team/{id}/players:
    get:
      tags:
        - Team
      summary: Getting all players in team
      operationId: getAllPlayersByTeam
      responses:
        "201":
          description: Successful retrieval of all players in team.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: integer
            minimum: 1
          in: query
          name: team
          required: true
          description: The team id
  /team/all/game/{name}:
    get:
      tags:
        - Team
      summary: Getting all teams by game
      operationId: getAllTeamsByGame
      responses:
        "201":
          description: Successful retrieval of all teams by game.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamDTO'

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      parameters:
        - schema:
            type: string
          in: query
          name: game
          required: true
          description: The game name
components:
  schemas:
    ErrorDto:
      title: Error
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    AccountDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        password:
          type: string
        isActive:
          type: boolean
    TeamDTO:
      properties:
        id:
          type: integer
        name:
          type: string
        captain:
          type: integer
        game:
          type: string
        isActive:
          type: boolean
    PlayerDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        password:
          type: string
        isActive:
          type: boolean
        team:
          type: integer
          required: false
        birthday:
          type: string
        countryOrigin:
          type: string
        playerName:
          type: string
          required: true
        image:
          type: string
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - OTHER
    UpdateAccountDTO:
      type: object
      properties:
        id:
          type: integer
          required: true
        name:
          type: string
          required: false
        email:
          type: string
          required: false
        password:
          type: string
          required: false
    CreateAccountDTO:
      type: object
      properties:
        name:
          type: string
          required: false
        email:
          type: string
          required: false
        password:
          type: string
          required: false
    UpdatePlayerDTO:
      type: object
      properties:
        id:
          type: integer
          required: true
        name:
          type: string
          required: false
        email:
          type: string
          required: false
        password:
          type: string
          required: false
        team:
          type: integer
          required: false
        birthday:
          type: string
          required: false
        countryOrigin:
          type: string
          required: false
        image:
          type: string
          required: false
        playerName:
          type: string
          required: true
        gender:
          type: string
          required: false
          enum:
            - MALE
            - FEMALE
            - OTHER
    CreatePlayerDTO:
      type: object
      properties:
        email:
          type: string
          required: true
        team:
          type: integer
          required: false
        birthday:
          type: string
          required: true
        countryOrigin:
          type: string
          required: true
        image:
          type: string
          required: true
        playerName:
          type: string
          required: true
        gender:
          type: string
          required: true
          enum:
            - MALE
            - FEMALE
            - OTHER
    UpdateTeamDTO:
      type: object
      properties:
        id:
          type: integer
          required: true
        name:
          type: string
          required: false
        captain:
          type: integer
          required: false
        game:
          type: string
          required: false
    CreateTeamDTO:
      type: object
      properties:
        name:
          type: string
          required: false
        captain:
          type: integer
          required: false
        game:
          type: string
          required: false